Introduced truncate option and enhanced checks
Introduced the option to truncate before load or to proceed with append only load.
Enhanced checks the data quality operator now recieves the check and the expected results and completes the check as passed in.

+++ plugins/operators/load_dimension.py
@@ -16,8 +16,9 @@ class LoadDimensionOperator(BaseOperator):
                  # Example:
                  # conn_id = your-connection-name
                  redshift_conn_id="redshift",
-                 table="songplays",
+                 table="",
                  insert_sql="",
+                 truncate_on=True,
                  *args, **kwargs):
 
         super(LoadDimensionOperator, self).__init__(*args, **kwargs)
@@ -26,18 +27,21 @@ class LoadDimensionOperator(BaseOperator):
         # self.conn_id = conn_id
         self.redshift_conn_id=redshift_conn_id
         self.table=table
+        self.truncate_on=truncate_on
         self.insert_sql=insert_sql
 
     """
         Operator execute to  clear data, and reload
+                 Default mode is delete and reload for dimension tables, unless truncate_on is False
     """
     def execute(self, context):
         # Redshift Hook
         redshift = PostgresHook(postgres_conn_id=self.redshift_conn_id)
         
-        # Clear data
-        self.log.info("Clearing data from destination Redshift dimension table {}".format(self.table))
-        redshift.run("DELETE FROM {}".format(self.table))
+        if self.truncate_on:
+            # Clear data
+            self.log.info("Clearing data from destination Redshift dimension table {}".format(self.table))
+            redshift.run("DELETE FROM {}".format(self.table))
         
         # Build Query
         insert_query="INSERT INTO {} ({})".format(self.table,self.insert_sql)
